designs

+++ Assets/Scripts/PointManager.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PointManager : MonoBehaviour
 {
-    public int GovernmentTolerance = 10;
+    public float GovernmentTolerance = 10;
     public int ResistancePoints = 0;
     public int Quota = 0;
     public int quotaTarget = 20;
